# parallelize the two slowest steps in the sync out