Part of our ongoing work to speed up the sync out.

On my local machine with 8 cores, this has a significant speedup. See these results:

### Before

```
                 user     system      total        real
rename       0.013404   0.004098   0.017502 (  0.019095)
restore      4.802004   0.328715 160.035807 (137.836081)
distribute 465.446141  14.723658 480.169799 (511.983090)
copy apps    0.029910   0.040080   0.069990 (  0.123882)
blockly      0.006614   0.000230  70.797939 ( 53.947998)
markdown     0.813416   0.076131   0.889547 (  0.889439)
hourofcode   2.553534   0.352502   2.906036 (  2.906916)
```

### After

```
                 user     system      total        real
rename       0.013888   0.003831   0.017719 (  0.019417)
restore      0.020309   0.100317 268.485802 ( 45.147419)
distribute  0.021499   0.108865 1497.125935 (297.122393)
copy apps    0.009859   0.039870   0.049729 (  0.049799)
blockly      0.006497   0.000547  69.314957 ( 52.540062)
markdown     1.073464   0.174946   1.248410 (  1.528806)
hourofcode   3.636642   0.358930   3.995572 (  4.024528)
```

The i18n-dev machine _currently_ only has two cores so I expect there to be less of a speedup there, but this opens us up to the possibility of speeding up the sync out by sizing up the ec2 instance.

Note that as a side effect, we lose the individual locale-by-locale progress reports. I think this is fine, as those resulted in more noise than I think we need. But! If we decide we want those back, we could achieve that by telling Parallel to use threads rather than processes.

## Testing story

Tested manually. Specifically, I ran the sync out both with and without the changes and compared the results to confirm that there is no difference other than in execution time.

# Reviewer Checklist:

- [ ] Tests provide adequate coverage
- [ ] Code is well-commented
- [ ] New features are translatable or updates will not break translations
- [ ] Relevant documentation has been added or updated
- [ ] User impact is well-understood and desirable
- [ ] Pull Request is labeled appropriately
- [ ] Follow-up work items (including potential tech debt) are tracked and linked